+++ src/components/AllProductList.js
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from 'react';
 import ProductCard from './ProductCard';
 import Pagination from './Pagination';
 import Loadingbar from './Loadingbar';
+import Dropdown from './Dropdown';
 import { getProducts } from '../services/api';
 
 const getPageSize = () => {
@@ -33,17 +34,7 @@ function AllProductList() {
   const [page, setPage] = useState(1);
   const [pageSize, setPageSize] = useState(getPageSize());
   const [totalPageNum, setTotalPageNum] = useState(0);
-  const [menuOpen, setMenuOpen] = useState(false);
   const [loading, setLoading] = useState(false);
-
-  const toggleMenu = () => {
-    setMenuOpen(prevState => !prevState);
-  };
-
-  const handleSortMenuClick = sortMenu => {
-    setMenuOpen(false);
-    setOrderBy(sortMenu.orderLabel);
-  };
 
   const handleLoad = async ({ orderBy, page, pageSize }) => {
     setLoading(true);
@@ -78,44 +69,34 @@ function AllProductList() {
     };
   }, [orderBy, page, pageSize]);
 
+  const ProductList =
+    products.length > 0 ? (
+      <ul className="product-list">
+        {products.map(product => (
+          <li key={product.id}>
+            <ProductCard product={product} />
+          </li>
+        ))}
+      </ul>
+    ) : (
+      <p className="list-none">상품이 없습니다.</p>
+    );
+
   return (
     <>
       <div className="section-header">
         <h3 className="title">판매 중인 상품</h3>
         <input type="text" className="search" placeholder="검색할 상품을 입력해주세요"></input>
-        <a href="./additem.html" className="btn-primary btn-sm btn-add" type="button">
+        <a href="./additem" className="btn-primary btn-sm btn-add" type="button">
           상품 등록하기
         </a>
-        <div className="dropdown">
-          <button onClick={toggleMenu} className="dropdown-toggle" type="button">
-            <span>{orderBy === 'recent' ? '최신순' : '좋아요순'}</span>
-          </button>
-          <ul className={`dropdown-menu ${menuOpen ? 'open' : ''}`}>
-            {Object.entries(SORT_MENU_INFO).map(([key, value]) => {
-              return (
-                <li key={key}>
-                  <button onClick={() => handleSortMenuClick(value)} className="dropdown-item">
-                    {value.label}
-                  </button>
-                </li>
-              );
-            })}
-          </ul>
-        </div>
+        <Dropdown
+          options={Object.values(SORT_MENU_INFO)}
+          selectedOption={SORT_MENU_INFO[orderBy.toUpperCase()]}
+          onSelect={option => setOrderBy(option.orderLabel)}
+        />
       </div>
-      {loading ? (
-        <Loadingbar />
-      ) : products.length > 0 ? (
-        <ul className="product-list">
-          {products.map(product => (
-            <li key={product.id}>
-              <ProductCard product={product} />
-            </li>
-          ))}
-        </ul>
-      ) : (
-        <p className="list-none">상품이 없습니다.</p>
-      )}
+      {loading ? <Loadingbar /> : ProductList}
       {totalPageNum > 1 && <Pagination totalPageNum={totalPageNum} activePageNum={page} onPageChange={onPageChange} />}
     </>
   );
